Consolidate e2e tests sample.spec.ts & try-fixture.sepc.ts
